api refactor, extra tests, fix interpretation
Moved api to Eval into an EvaluationTree object which has
getinterpretation and evaluate. Abstracted this logic in the mathparser
too.

Added a couple of extra tests, and updated tests to use the new API.

Fixed some interpretation issues with binary, decimal, hex, and octal
input.

Added " and ' as unit types for inches and feet.